decouple ScriptSpec into HashSpec and ValidationSpec
